feat: reinstall transient `lean4-pr-releases` toolchains
This PR adds the option to check if the current toolchain is a transient one hosted on the lean4-pr-releases repository. If so, `lean-action` will remove the old version of the toolchain. This ensures a fresh version of the toolchain is always used to run CI.

This is an important step in many Mathlib CI workflows, and adding it to this action would allow us to switch some custom workflows over to run mostly via `lean-action`.